Working, but inefficient genetic algorithm
It approaches the lowest energy, and the best weights. We would expect that all of the weight would go to the most negative energy term. This is indeed what we observe, at the bottom of this notebook. 

Genetic algorithm is still inefficient, as it is lacking several subroutines (i.e. crossover, mutation, tournament selection, etc.).